Backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pandas
numpy
scikit-learn==1.6.1
matplotlib
statsmodels
joblib
pymongo[srv]
motor
python-jose[cryptography]
passlib[bcrypt]
bcrypt
email-validator
stripe
python-dotenv
httpx